# openapi.yaml
openapi: 3.0.0
info:
  title: Notification Server HTTP API
  description: |
    HTTP API for managing notifications, channels, and client access.
    This API handles client management, channel creation, and access control.
  version: 1.0.0
  contact:
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
components:
  schemas:
    Notification:
      type: object
      required:
        - channel
        - message
      properties:
        channel:
          type: string
          description: Channel name
        message:
          type: string
          description: Notification message content
        timestamp:
          type: string
          format: date-time
          description: Message timestamp
        metadata:
          type: object
          description: Additional message metadata
          properties:
            priority:
              type: string
              enum: [low, medium, high]
              description: Message priority
            tags:
              type: array
              items:
                type: string
              description: Message tags
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        code:
          type: integer
          description: Error code
        details:
          type: object
          description: Additional error details
...
    ClientMetadata:
      type: object
      properties:
        userAgent:
          type: string
          description: Client user agent
        environment:
          type: string
          description: Client environment
        custom:
          type: object
          description: Custom metadata fields
paths:
  /api/clients:
    post:
      summary: Generate or validate a client ID
      description: |
        Generates a new client ID or validates an existing one.
        If a clientId is provided in the request, it will be validated.
        If no clientId is provided or it's invalid, a new one will be generated.
      operationId: generateOrValidateClientId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clientId:
                  type: string
                  description: Optional client ID to validate
                metadata:
                  $ref: '#/components/schemas/ClientMetadata'
      responses:
        '200':
          description: Client ID is valid
          content:
            application/json:
              schema:
                type: object
                required:
                  - clientId
                  - message
                properties:
                  clientId:
                    type: string
                    description: Validated client ID
                  message:
                    type: string
                    description: Status message
        '201':
          description: New client ID generated
          content:
            application/json:
              schema:
                type: object
                required:
                  - clientId
                  - message
                properties:
                  clientId:
                    type: string
                    description: Generated client ID
                  message:
                    type: string
                    description: Status message
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/clients/{clientId}:
    get:
      summary: Validate client ID
      description: Checks if a client ID is valid
      operationId: validateClientId
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
          description: Client ID to validate
      responses:
        '200':
          description: Client ID is valid
          content:
            application/json:
              schema:
                type: object
                required:
                  - valid
                properties:
                  valid:
                    type: boolean
                    description: Whether the client ID is valid
        '404':
          description: Client ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a client
      description: Deletes a client and all its associated data (subscriptions, channel access)
      operationId: deleteClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
          description: Client ID to delete
      responses:
        '200':
          description: Client deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
                  clientId:
                    type: string
                    description: Deleted client ID
                  deletedSubscriptions:
                    type: integer
                    description: Number of subscriptions removed
                  deletedChannels:
                    type: integer
                    description: Number of channel access entries removed
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/channels:
    post:
      summary: Create a new channel
      description: Creates a new channel with specified rules. Returns 409 if channel already exists.
      tags:
        - Channel Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - channel
              properties:
                channel:
                  type: string
                  description: Name of the channel to create
                rules:
                  type: object
                  description: Channel access rules
                  properties:
                    isPublic:
                      type: boolean
                      description: Whether the channel is public
                    allowedClientIds:
                      type: array
                      items:
                        type: string
                      description: List of allowed client IDs
                    maxSubscribers:
                      type: integer
                      description: Maximum number of subscribers allowed
      responses:
        '201':
          description: Channel created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  channel:
                    type: string
                  rules:
                    type: object
        '400':
          description: Bad request - channel name is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Channel already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a channel
      description: Deletes an existing channel and all its subscriptions
      tags:
        - Channel Management
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
          description: Name of the channel to delete
      responses:
        '200':
          description: Channel deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  channel:
                    type: string
        '400':
          description: Bad request - channel name is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/channels/{channel}/access/{clientId}:
    post:
      summary: Grant channel access
      description: Grants a client access to a channel
      operationId: grantChannelAccess
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
          description: Channel name
        - name: clientId
          in: path
          required: true
          schema:
            type: string
          description: Client ID to grant access to
      responses:
        '200':
          description: Access granted successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    description: Whether access was granted
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Channel or client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Revoke channel access
      description: Revokes a client's access to a channel
      operationId: revokeChannelAccess
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
          description: Channel name
        - name: clientId
          in: path
          required: true
          schema:
            type: string
          description: Client ID to revoke access from
      responses:
        '200':
          description: Access revoked successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    description: Whether access was revoked
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Channel or client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/clients/{clientId}/channels:
    get:
      summary: Get accessible channels
      description: Retrieves all channels a client has access to
      operationId: getClientChannels
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
          description: Client ID
      responses:
        '200':
          description: List of accessible channels
          content:
            application/json:
              schema:
                type: object
                required:
                  - channels
                properties:
                  channels:
                    type: array
                    items:
                      type: string
                    description: List of channel names
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications:
    post:
      summary: Publish a notification
      description: Publishes a notification to a channel
      operationId: publishNotification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - channel
                - message
              properties:
                channel:
                  type: string
                  description: Channel to publish to
                message:
                  type: string
                  description: Notification message
                metadata:
                  type: object
                  description: Additional message metadata
      responses:
        '200':
          description: Notification published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/{channel}:
    get:
      summary: Get notification history
      description: Retrieves the notification history for a channel
      operationId: getNotificationHistory
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
          description: Channel name
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Maximum number of notifications to return
      responses:
        '200':
          description: Notification history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/clients/{clientId}/subscriptions:
    get:
      summary: Get client subscriptions
      description: Retrieves all channels a client is subscribed to
      operationId: getClientSubscriptions
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
          description: Client ID
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSubscriptions'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...